Add files via upload
Software Engineer changes from semester to semester, so I am safe to upload this whole document. My group made a Virtual Machine in C# and I wrote the backend, another student wrote the front, another wrote a controller between them and the fourth documented the process and formed a presentation on it. It turns out that C# is really nice to use.